backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# API Documentation
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1

# Caching & Performance
redis==5.0.1
django-redis==5.4.0
django-cacheops==6.1

# Real-time Features
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# File Storage & Media
Pillow==10.1.0
boto3==1.34.0
django-storages==1.14.2

# Payment Processing
stripe==7.8.0

# Email & Notifications
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
sendgrid==6.10.0

# Monitoring & Logging
sentry-sdk==1.38.0
django-extensions==3.2.3

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
